copaco-azure-function/requirements.txt
azure.functions
azure-functions-worker
requests
pydantic
httpx
